+++ src/main/java/com/pax/market/api/sdk/java/api/factoryModel/dto/FactoryPageResponse.java
@@ -1,0 +1,12 @@
+package com.pax.market.api.sdk.java.api.factoryModel.dto;
+
+import com.pax.market.api.sdk.java.api.base.dto.PageResponse;
+
+/**
+ * @author shifan
+ * @date 2025/2/17
+ */
+public class FactoryPageResponse extends PageResponse<FactoryDTO> {
+
+
+}